# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: opengym API
  version: 1.0.0
  description: API for opengym
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/auth/me:
    get:
      summary: Get authenticated user
      description: Returns the currently authenticated user
      tags:
        - Authentication
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized - invalid or missing token
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      required:
        - id
        - email
        - isDemo
      properties:
        id:
          type: string
          description: Unique user identifier
          example: "12345"
        email:
          type: string
          format: email
          description: User's email address
          example: user@example.com
        name:
          type: string
          description: User's full name
          example: John Doe
        picture:
          type: string
          format: uri
          description: URL to user's profile picture
          example: https://example.com/avatar.jpg
        createdAt:
          type: string
          format: date-time
          description: Timestamp when user was created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when user was last updated
        isDemo:
          type: boolean
          description: Whether the user is a demo user
          example: false
...
    Error:
      type: string
      description: Error message string